requirements.txt
great-expectations>=0.18.0
pyspark>=3.4.0
apache-airflow>=2.8.0
apache-airflow-providers-apache-spark>=4.7.0
boto3>=1.28.0
